Add some logging to each step

+++ desktopgnome.sh
@@ -4,19 +4,25 @@
 
 # Disable read-only status
 if test (sudo steamos-readonly status) != "disabled" 
+    echo "Making SteamOS writable for upgrades"
     sudo steamos-readonly disable
 end
 
 # Prepare system for the pacman repos
+echo "Preparing system for upgrade"
 sudo pacman-key --init
 sudo pacman-key --populate archlinux
 sudo pacman-key --populate holo
 
 # Install GNOME packages
+echo "Installing GNOME"
 sudo pacman --overwrite "*" -Syu gnome-keyring gnome-screenshot gnome-shell-extensions gnome-shell gnome-control-center gnome-tweaks gnome-software nautilus sushi xdg-user-dirs-gtk xdg-desktop-portal-gnome xorg-xinput
 
 # Backup startplasma-x11
-sudo mv --no-clobber /usr/bin/startplasma-x11 /usr/bin/startplasma-x11.bak
+echo "Backing up KDE Plasma"
+if not test -e /usr/bin/startplasma-x11.bak
+    sudo mv --no-clobber /usr/bin/startplasma-x11 /usr/bin/startplasma-x11.bak
+end
 
 # Create startgnome-x11
 echo "\
@@ -25,16 +31,19 @@ export XDG_SESSION_TYPE=x11
 export GDK_BACKEND=x11
 export XDG_CURRENT_DESKTOP=GNOME
 exec gnome-session
-" | sudo tee /usr/bin/startgnome-x11
+" | sudo tee /usr/bin/startgnome-x11 > /dev/null
 sudo chmod 755 /usr/bin/startgnome-x11
 
 # Replace startplasma-x11 with startgnome-x11
+echo "Replacing KDE Plasma with GNOME"
 sudo ln -s /usr/bin/startgnome-x11 /usr/bin/startplasma-x11
 
 # Disable HiDPI 2x scaling
+echo "Fixing UI scaling"
 gsettings set org.gnome.desktop.interface scaling-factor 1
 
 # Autostart Steam silently on GNOME
+echo "Setting Steam to autostart silently on GNOME"
 echo "\
 [Desktop Entry]
 Name=Steam
@@ -44,5 +53,5 @@ Terminal=false
 Type=Application
 PrefersNonDefaultGPU=true
 X-KDE-RunOnDiscreteGpu=true
-" | tee ~/.config/autostart/steam.desktop
+" > ~/.config/autostart/steam.desktop
 chmod 644 ~/.config/autostart/steam.desktop
